Reformatted OrderForm
