+++ src/InventoryTable.tsx
@@ -1,0 +1,5 @@
+function Message(){
+    return <h1>Hello World</h1>;
+}
+
+export default Message;
